requirements.txt
Django=4.2.1    
djangorestframework=3.14.0   